Chore: clean up error handling in AutoEditor
