+++ src/components/CombinedFeaturePanel/editors/WheelchairEditor.tsx
@@ -3,10 +3,10 @@ import { t } from "@transifex/native";
 import type React from "react";
 import { useEffect, useState } from "react";
 import { PrimaryButton, SecondaryButton } from "~/components/shared/Buttons";
+import { unknownCategory } from "~/domains/categories/functions/cache";
 import { useCurrentLanguageTagStrings } from "~/lib/context/LanguageTagContext";
 import type { YesNoLimitedUnknown } from "~/lib/model/ac/Feature";
 import { isWheelchairAccessible } from "~/lib/model/accessibility/isWheelchairAccessible";
-import { unknownCategory } from "~/domains/categories/functions/cache";
 import { AccessibilityView } from "~/pages/[placeType]/[id]/report/send-report-to-ac";
 import Icon from "../../shared/Icon";
 import { StyledReportView } from "../ReportView";
@@ -93,8 +93,9 @@ export const WheelchairEditor: React.FC<BaseEditorProps> = ({
                 }
                 valueName="Partially"
               >
-                {t`Entrance has one step with max. 3 inches height, most rooms are
-                without steps`}
+                {t(
+                  "Entrance has one step with max. 3 inches height, most rooms are without steps",
+                )}
               </AccessibilityView>
 
               <AccessibilityView
@@ -108,8 +109,9 @@ export const WheelchairEditor: React.FC<BaseEditorProps> = ({
                 icon={<Icon size="medium" accessibility="no" category={cat} />}
                 valueName="Not at all"
               >
-                {t`Entrance has a high step or several steps, none of the rooms are
-                accessible.`}
+                {t(
+                  "Entrance has a high step or several steps, none of the rooms are accessible.",
+                )}
               </AccessibilityView>
             </form>
