+++ src/lib/fetchers/osm-api/getOsmParametersFromFeature.ts
@@ -1,13 +1,14 @@
-import { getFeatureId } from "../../model/ac/Feature";
 import { getOSMType } from "../../model/osm/generateOsmUrls";
 
 export default function getOsmParametersFromFeature(feature, tagKey) {
   const tagName = typeof tagKey === "string" ? tagKey : tagKey[0];
   const osmType = getOSMType(feature);
-  const osmId = feature.id;
+  const prefixedId = feature._id;
+  const osmId = prefixedId.replace(/^(?:[^/]+\/)?/, "");
   return {
     tagName,
     osmType,
+    prefixedId,
     osmId,
   };
 }
